# public/docs/openapi.yaml
openapi: 3.0.0
info:
  title: 'Cairo Store API'
  description: "كل نقاط المتجر ولوحة التحكم.\r\n\r\nملاحظة عامة على الأخطاء: نقاط JSON في هذا المشروع تُرجع 200 حتى عند\r\nفشل التحقق، والنتيجة تُقرأ من الحقل success في الجسم لا من كود HTTP.\r\nالاستثناءات الوحيدة هي 302 (تحويل لتسجيل الدخول) و403 (صلاحية ناقصة)\r\nو404 (راوت غير موجود). هذا سلوك قائم يعتمد عليه الـfrontend، ووُثّق\r\nهنا كما هو لا كما ينبغي أن يكون."
  version: 1.1.0
servers:
  -
    url: 'http://localhost/STORE/public'
    description: 'Local / Development Server'
paths:
...
  /checkout:
    get:
      tags:
        - 'Store - Checkout'
      summary: 'Checkout page (three steps)'
      operationId: 3c9e7f1f4ad315498f6c07de22b0082f
      responses:
        '200':
          description: 'HTML page'
        '302':
          description: 'Redirect home with the login modal opened when not signed in'
      security:
        -
          userSessionAuth: []
    post:
      tags:
        - 'Store - Checkout'
      summary: 'Create the order from the cart contents'
      description: 'The client sends **what and how many** and nothing else. Every monetary value is read from the database inside the same transaction that reserves stock. The shown_price field is compared and never stored: if it differs from the database price the whole order is refused and the server answers error_code=price_changed together with the correct prices.'
      operationId: b9a528bfe771ba04bed8dc5cf05bef51
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required:
                - items
                - address_id
                - idempotency_key
                - csrf_token
              properties:
                items:
                  type: array
                  items: { required: [product_id, qty], properties: { product_id: { type: integer }, variant_id: { type: integer, nullable: true }, qty: { type: integer, maximum: 100, minimum: 1 }, shown_price: { description: 'The price the customer was shown — for comparison only; never stored and never computed from', type: number, format: float } }, type: object }
                address_id:
                  type: integer
                payment_method:
                  type: string
                  default: cash_on_delivery
                idempotency_key:
                  type: string
                csrf_token:
                  type: string
              type: object
      responses:
        '200':
          description: 'Operation result. Success carries order_id and redirect. Failure carries an error_code out of: price_changed · unavailable · out_of_stock · error · csrf_invalid. Only price_changed additionally carries items with the correct prices.'
          content:
            application/json:
              schema:
                oneOf:
                  - { $ref: '#/components/schemas/ApiResponse' }
                  - { $ref: '#/components/schemas/ApiError' }
      security:
        -
          userSessionAuth: []
  /checkout/cancel-order:
    post:
      tags:
        - 'Store - Checkout'
      summary: 'Cancel an order, by the user'
      description: 'Cancelling is allowed only before an admin takes the order. The same shared button in views/shared/order-cancel-button.php serves the admin interface through a different endpoint.'
      operationId: 97b7adab83828a0e8f3ee5568419e90a
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              required:
                - order_id
                - csrf_token
              properties:
                order_id:
                  type: integer
                csrf_token:
                  type: string
              type: object
      responses:
        '200':
          description: 'Operation result. The success field separates success from failure — the HTTP status stays 200 either way. On CSRF failure the body carries error_code=csrf_invalid.'
          content:
            application/json:
              schema:
                oneOf:
                  - { $ref: '#/components/schemas/ApiResponse' }
                  - { $ref: '#/components/schemas/ApiError' }
      security:
        -
          userSessionAuth: []
  /checkout/confirmation:
    get:
      tags:
        - 'Store - Checkout'
      summary: 'Order confirmation page, after checkout'
      operationId: 700822d16d5b6157d2dde3a3b9f3b073
      parameters:
        -
          name: id
          in: query
          schema:
            type: integer
      responses:
        '200':
          $ref: '#/components/responses/HtmlPage'
        '404':
          $ref: '#/components/responses/NotFoundPage'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
      security:
        -
          userSessionAuth: []
...
components:
  schemas:
    ApiResponse:
      title: 'غلاف استجابة JSON الموحّد'
      description: "الشكل الذي تُرجعه كل نقطة JSON في المشروع، من Controller::respond().\r\n\r\nكود HTTP يبقى 200 حتى عند الفشل. النتيجة تُقرأ من الحقل success لا\r\nمن كود الحالة. هذا سلوك قائم يعتمد عليه الـfrontend في 34 ملف JS."
      required:
        - success
        - message
      properties:
        success:
          description: 'مصدر الحقيقة الوحيد لنجاح العملية أو فشلها.'
          type: boolean
          example: true
        message:
          description: 'نصّ للعرض على المستخدم. لا تعتمد عليه برمجياً — راجع error_code.'
          type: string
          example: 'تمت العملية بنجاح.'
      type: object
    ApiError:
      title: 'استجابة فشل تحمل رمزاً صريحاً'
      description: "مثل ApiResponse لكن success=false ومعها error_code اختياري.\r\n\r\nالرمز عقد بين الخادم والمتصفح، والنصّ للعرض وحده. الفصل بينهما ليس\r\nتفضيلاً أسلوبياً: كان js/core/csrf.js يكتشف فشل CSRF بمطابقة بداية\r\nنصّ الرسالة، فأي نقطة تصوغ رسالتها بشكل آخر تفقد إعادة المحاولة\r\nالتلقائية بصمت. حدث ذلك ثلاث مرّات قبل أن يُستبدل النصّ برمز."
      required:
        - success
        - message
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: 'Invalid CSRF token, please refresh and try again.'
        error_code:
          description: 'رمز ثابت تقرأه الآلة. يُضاف حين يحتاج العميل التصرّف لا العرض فقط.'
          type: string
          example: csrf_invalid
          enum:
            - csrf_invalid
      type: object
...
  responses:
...
    HtmlPage:
      description: 'صفحة HTML كاملة.'
      content:
        text/html: {  }
...
    NotFoundPage:
      description: 'راوت غير مسجَّل أو مورد غير موجود. صفحة HTML من ErrorPage::notFound.'
      content:
        text/html: {  }
    ServiceUnavailable:
      description: "عطل تقني يمنع إكمال الطلب — أبرزه فشل الاتصال بقاعدة البيانات.\r\n\r\n503 لا 500 عمداً: الخدمة غير متاحة مؤقتاً لا «خطأ في الخادم»،\r\nوالفرق يهمّ محرّكات البحث وأدوات المراقبة. التفاصيل تذهب إلى سجلّ\r\nالأخطاء ولا تُطبع للزائر أبداً — رسالة PDO تحمل اسم المضيف واسم\r\nالقاعدة واسم المستخدم."
      content:
        text/html: {  }
...
  securitySchemes:
    adminSessionAuth:
      type: apiKey
      description: 'PHP session cookie — يُنشأ تلقائياً عند تسجيل الدخول ويُرسل مع كل طلب محمي'
      name: admin_session
      in: cookie
    userSessionAuth:
      type: apiKey
      description: "جلسة المستخدم العادي — منفصلة تماماً عن admin_session بالاسم\r\nوالمحتوى، فلا يمكن الوصول لبيانات الأدمن من جلسة مستخدم ولا العكس\r\n(راجع isUser/isAdmin في auth_helper.php)."
      name: PHPSESSID
      in: cookie
...